@@ -1,24 +1,23 @@
 {
   "name": "demo",
   "version": "0.1.6",
+  "type": "module",
   "scripts": {
     "dev": "vite",
     "build": "vite build",
     "serve": "vite preview"
   },
   "devDependencies": {
-    "autoprefixer": "^10.4.16",
-    "eslint": "^8.51.0",
-    "eslint-config-lxsmnsyc": "^0.6.5",
-    "postcss": "^8.4.31",
-    "typescript": "^5.2.2",
-    "vite": "^4.5.0",
-    "vite-plugin-solid": "^2.7.2"
+    "autoprefixer": "^10.4.17",
+    "postcss": "^8.4.33",
+    "typescript": "^5.3.3",
+    "vite": "^5.0.12",
+    "vite-plugin-solid": "^2.9.1"
   },
   "dependencies": {
-    "solid-js": "^1.8.1",
+    "solid-js": "^1.8.12",
     "solid-mason": "0.1.6",
-    "tailwindcss": "^3.3.3"
+    "tailwindcss": "^3.4.1"
   },
   "private": true,
   "publishConfig": {
