@@ -14,16 +14,14 @@
     "pridepack"
   ],
   "devDependencies": {
-    "@types/node": "^20.8.7",
-    "eslint": "^8.51.0",
-    "eslint-config-lxsmnsyc": "^0.6.5",
-    "pridepack": "2.5.1",
-    "solid-js": "^1.8.1",
+    "@types/node": "^20.11.5",
+    "pridepack": "2.6.0",
+    "solid-js": "^1.8.12",
     "tslib": "^2.6.2",
-    "typescript": "^5.2.2"
+    "typescript": "^5.3.3"
   },
   "peerDependencies": {
-    "solid-js": "^1.6"
+    "solid-js": "^1.8"
   },
   "scripts": {
     "prepublishOnly": "pridepack clean && pridepack build",
